ta: crypto_perf: update TEE_Param according to GP131
Updates TEE_Param according to TEE Internal Core API version 1.3.1.

+++ ta/crypto_perf/ta_alg_asymcipher.c
@@ -114,8 +114,8 @@ TEE_Result TA_AsymCipherProcessAlgo(uint32_t algo __unused, TEE_Param params[4])
 	/* Asymmetric Cipher Process: Don't need to check again input params */
 	TEE_Result res;
 	void *in, *out;
-	uint32_t inSize;
-	uint32_t outSize;
+	size_t inSize;
+	size_t outSize;
 
 	in      = params[0].memref.buffer;
 	inSize  = params[0].memref.size;
